+++ Configuration Management/SettingsWindow.xaml.cs
@@ -851,6 +851,7 @@ namespace Configuration_Management
             "ServerBase" => "Column.ServerBase",
             "LastLaunch" => "Column.LastLaunch",
             "Size" => "Column.Size",
+            "Actions" => "Column.Actions",
             _ => "Column.Name"
         });